Revert "Upconvert various types to int."
This reverts commit 83ad33a8aed1fb00e422b6abd33c3e8942821c24.

BUG=499241
TBR=hlundin

Review URL: https://codereview.webrtc.org/1179953003

Cr-Commit-Position: refs/heads/master@{#9418}

+++ webrtc/modules/audio_coding/codecs/g722/g722_interface.c
@@ -39,7 +39,7 @@ int16_t WebRtcG722_EncoderInit(G722EncInst *G722enc_inst)
     }
 }
 
-int WebRtcG722_FreeEncoder(G722EncInst *G722enc_inst)
+int16_t WebRtcG722_FreeEncoder(G722EncInst *G722enc_inst)
 {
     // Free encoder memory
     return WebRtc_g722_encode_release((G722EncoderState*) G722enc_inst);
@@ -79,7 +79,7 @@ int16_t WebRtcG722_DecoderInit(G722DecInst *G722dec_inst)
     }
 }
 
-int WebRtcG722_FreeDecoder(G722DecInst *G722dec_inst)
+int16_t WebRtcG722_FreeDecoder(G722DecInst *G722dec_inst)
 {
     // Free encoder memory
     return WebRtc_g722_decode_release((G722DecoderState*) G722dec_inst);
